Refactor sitemap
